# swagger.yaml
openapi: 3.0.3
info:
  title: Frames API
  description: API for managing picture frames, walls, and 3D models
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    Picture:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        wall_id:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
        image_path:
          type: string
        original_image_path:
          type: string
        thumbnail_path:
          type: string
          nullable: true
        width_px:
          type: integer
        height_px:
          type: integer
        created_at:
          type: string
          format: date-time
        frames:
          type: array
          items:
            $ref: '#/components/schemas/PictureFrame'
    PictureFrame:
      type: object
      properties:
        id:
          type: integer
        picture_id:
          type: integer
        name:
          type: string
        width_cm:
          type: number
          format: float
        height_cm:
          type: number
          format: float
        depth_cm:
          type: number
          format: float
        width_inches:
          type: number
          format: float
        height_inches:
          type: number
          format: float
        depth_inches:
          type: number
          format: float
        frame_color:
          type: string
        frame_material:
          type: string
        mat_width_inches:
          type: number
          format: float
        mat_color:
          type: string
        model_path:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - BearerAuth: []
paths:
...
  /pictures:
    get:
      tags:
        - Pictures
      summary: Get all pictures
      description: Retrieve all pictures for the authenticated user
      responses:
        '200':
          description: List of pictures
          content:
            application/json:
              schema:
                type: object
                properties:
                  pictures:
                    type: array
                    items:
                      $ref: '#/components/schemas/Picture'

    post:
      tags:
        - Pictures
      summary: Create a new picture
      description: Upload and create a new picture from an image file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Picture image file (png, jpg, jpeg, gif, webp)
                name:
                  type: string
                  example: Artwork Title
                description:
                  type: string
                  example: Description of the artwork
                wall_id:
                  type: integer
                  description: Optional wall assignment
      responses:
        '201':
          description: Picture created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  picture:
                    $ref: '#/components/schemas/Picture'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pictures/{picture_id}:
    get:
      tags:
        - Pictures
      summary: Get a specific picture
      description: Retrieve a picture by ID
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Picture details
          content:
            application/json:
              schema:
                type: object
                properties:
                  picture:
                    $ref: '#/components/schemas/Picture'
        '404':
          description: Picture not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Pictures
      summary: Update a picture
      description: Update picture details
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                wall_id:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Picture updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  picture:
                    $ref: '#/components/schemas/Picture'
        '404':
          description: Picture not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Pictures
      summary: Delete a picture
      description: Delete a picture and all associated frames
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Picture deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Picture not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pictures/{picture_id}/image:
    put:
      tags:
        - Pictures
      summary: Update picture image
      description: Update a picture's image (for recrop functionality)
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  picture:
                    $ref: '#/components/schemas/Picture'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Picture not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pictures/{picture_id}/frames:
    post:
      tags:
        - Frames
      summary: Create a frame
      description: Create a 3D frame for a picture with specified dimensions
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - width
                - height
              properties:
                name:
                  type: string
                  example: 8x10 Frame
                unit:
                  type: string
                  enum: [inches, cm]
                  default: inches
                  example: inches
                width:
                  type: number
                  format: float
                  example: 8.0
                height:
                  type: number
                  format: float
                  example: 10.0
                depth:
                  type: number
                  format: float
                  example: 1.0
                frame_color:
                  type: string
                  example: "#8B4513"
                frame_material:
                  type: string
                  example: wood
                mat_width:
                  type: number
                  format: float
                  example: 0.0
                mat_color:
                  type: string
                  example: "#FFFFFF"
      responses:
        '201':
          description: Frame created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  frame:
                    $ref: '#/components/schemas/PictureFrame'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Picture not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pictures/{picture_id}/frames/{frame_id}:
    put:
      tags:
        - Frames
      summary: Update a frame
      description: Update frame dimensions and styling
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
        - name: frame_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                unit:
                  type: string
                  enum: [inches, cm]
                width:
                  type: number
                  format: float
                height:
                  type: number
                  format: float
                depth:
                  type: number
                  format: float
                frame_color:
                  type: string
                frame_material:
                  type: string
      responses:
        '200':
          description: Frame updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  frame:
                    $ref: '#/components/schemas/PictureFrame'
        '404':
          description: Picture or frame not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Frames
      summary: Delete a frame
      description: Delete a frame and its model file
      parameters:
        - name: picture_id
          in: path
          required: true
          schema:
            type: integer
        - name: frame_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Frame deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Picture or frame not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...